fix: staker coin ages calc

+++ contracts/interfaces/IUserManager.sol
@@ -8,27 +8,9 @@ pragma solidity 0.8.16;
 interface IUserManager {
     function memberFrozen(address staker) external view returns (uint256);
 
-    function stakers(address staker)
-        external
-        view
-        returns (
-            bool,
-            uint96,
-            uint96,
-            uint64,
-            uint256,
-            uint256
-        );
+    function stakers(address staker) external view returns (bool, uint96, uint96, uint64, uint256, uint256);
 
-    function vouchers(address borrower, uint256 index)
-        external
-        view
-        returns (
-            address,
-            uint96,
-            uint96,
-            uint64
-        );
+    function vouchers(address borrower, uint256 index) external view returns (address, uint96, uint96, uint64);
 
     function vouchees(address staker, uint256 index) external view returns (address, uint96);
 
@@ -65,11 +47,7 @@ interface IUserManager {
 
     function withdrawRewards() external;
 
-    function debtWriteOff(
-        address staker,
-        address borrower,
-        uint96 amount
-    ) external;
+    function debtWriteOff(address staker, address borrower, uint256 amount) external;
 
     /**
      *  @dev Check if the account is a valid member
@@ -132,14 +110,7 @@ interface IUserManager {
      *  @return  user's effective staked amount
      *           user's effective locked amount
      */
-    function getStakeInfo(address staker, uint256 pastBlocks)
-        external
-        view
-        returns (
-            uint256,
-            uint256,
-            bool
-        );
+    function getStakeInfo(address staker, uint256 pastBlocks) external view returns (uint256, uint256, bool);
 
     /**
      * @dev Update the frozen info by the comptroller
@@ -148,13 +119,7 @@ interface IUserManager {
      * @return  effectStaked user's total stake - frozen
      *          effectLocked user's locked amount - frozen
      */
-    function onWithdrawRewards(address staker, uint256 pastBlocks)
-        external
-        returns (
-            uint256,
-            uint256,
-            bool
-        );
+    function onWithdrawRewards(address staker, uint256 pastBlocks) external returns (uint256, uint256, bool);
 
     /**
      * @dev Update the frozen info by the utoken repay
@@ -168,11 +133,7 @@ interface IUserManager {
      *  @param amount Borrow or repay amount(Including previously accrued interest)
      *  @param isBorrow True is borrow, false is repay
      */
-    function updateLocked(
-        address borrower,
-        uint96 amount,
-        bool isBorrow
-    ) external;
+    function updateLocked(address borrower, uint96 amount, bool isBorrow) external;
 
     /**
      *  @dev Get the user's deposited stake amount
